refactor: split ios.rs into platform/ios/ module (SRP)
